[BUGFIX] Update mandatory icon if field is cleared via button (#7)
Metadata fields can have a colored icon that indicates if a value
is mandatory. The icon color changes if a value is set or removed
by manual input.

Now, the color is also updated when a user uses the clear button
of the field ("close" icon on the right).

(cherry picked from commit ae87f02f89d66904788fd658ae055c930fc3da8e)