[TASK] Improve icon rendering on mandatory fields
Mandatory fields are indicated with a red colored icon
(or yellow/orange, if it is mandatory in certain scenarios).
When a mandatory field is filled, the icon color
switches to green.

To improve usability, we now also change the icon from
an exclamation mark (in a triangle) to a checkmark.

Colors and icon visibility are now set with a new backend
stylesheet, which also support the color modes in TYPO3 v13.
Icon colors pass WCAG AA in both color modes.